bug fix progress

+++ CheckpointManager.cpp
@@ -15,7 +15,7 @@ RunCheckpoint* CheckpointManager::undo(){
     }
 }
 CheckpointManager::~CheckpointManager(){
-    for(int i = 0; i < history.size(); i++){
+    for(size_t i = 0; i < history.size(); i++){
         delete history[i];
     }
 }
